src/test/java/kr/ac/jejunu/user/UserDaoTests.java
package kr.ac.jejunu.user;

import org.junit.jupiter.api.Test;

import java.sql.SQLException;

import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.greaterThan;
import static org.hamcrest.core.Is.is;
public class UserDaoTests {
    @Test
    public void get() throws SQLException, ClassNotFoundException {
        Long id = 1l;
        String name = "haeeun";
        String password = "겟1234";
        UserDao userDao = new JejuUserDao();
        User user = userDao.findId(id);
        assertThat(user.getId(), is(id));
        assertThat(user.getName(), is(name));
        assertThat(user.getPassword(), is(password));
    }

    @Test
    public void insert() throws SQLException, ClassNotFoundException {
        User user = new User();

        String name = "haeeun";
        String password = "insert1234";
        user.setName(name);
        user.setPassword(password);

        UserDao userDao = new JejuUserDao();
        userDao.insert(user);
        assertThat(user.getId(), greaterThan(1l));

        User insertedUser = userDao.findId(user.getId());
        assertThat(user.getName(), is(name));
        assertThat(user.getPassword(),is(password));
    }

    @Test
    public void getForHalla() throws SQLException, ClassNotFoundException {
        Long id = 1l;
        String name = "haeeun";
        String password = "겟1234";
        UserDao userDao = new HallaUserDao();
        User user = userDao.findId(id);
        assertThat(user.getId(), is(id));
        assertThat(user.getName(), is(name));
        assertThat(user.getPassword(), is(password));
    }

    @Test
    public void insertForHalla() throws SQLException, ClassNotFoundException {
        User user = new User();

        String name = "haeeun";
        String password = "insert1234";
        user.setName(name);
        user.setPassword(password);

        UserDao userDao = new HallaUserDao();
        userDao.insert(user);
        assertThat(user.getId(), greaterThan(1l));

        User insertedUser = userDao.findId(user.getId());
        assertThat(user.getName(), is(name));
        assertThat(user.getPassword(),is(password));
    }
}
